clearing and some shit

+++ src/otherFunctions.js
@@ -1,21 +1,25 @@
 import { deleteDoc, doc } from "firebase/firestore";
 import { deleteObject, ref } from "firebase/storage";
-import { auth, db, storage } from "./firebaseConfig";
+import { db, storage } from "./firebaseConfig";
 
+
+export const logError = (...e) =>{
+  console.log(...e);
+}
 export const deleteAllParkings = async (parkings) => {
+  if(!parkings)return;
   parkings.forEach((parking) => {
     deleteDoc(doc(db, "parkings", parking))
-      .then((e) => console.log("file success", e))
-      .catch((e) => console.log("file failed", e));
+      .then(() => {})
+      .catch((e) => logError("file failed", e));
     deleteObject(ref(storage, "parkings" + parking))
-      .then((e) => console.log("image success", e))
-      .catch((e) => console.log("image failed", e));
+      .then(() => {})
+      .catch((e) => logError("image failed", e));
   });
 };
 
 export const isAvailable = (parking) => {
   if (!parking.availability) return null;
-  console.log(parking.availability.monday);
   let t = new Date();
   let weekDays = [
     "Sunday",
@@ -31,7 +35,6 @@ export const isAvailable = (parking) => {
     String(t.getHours()).padStart(2, "0") +
     ":" +
     String(t.getMinutes()).padStart(2, "0");
-  console.log(day, time);
   return (
     (parking.availability[day][0]?.start < time &&
       time < parking.availability[day][0]?.end) ||
